Improve destroy command

+++ cmd/mantil/cmd/destroy.go
@@ -1,11 +1,13 @@
 package cmd
 
 import (
+	"fmt"
 	"log"
 	"os"
 
 	"github.com/atoz-technology/mantil-cli/internal/aws"
 	"github.com/atoz-technology/mantil-cli/internal/github"
+	"github.com/atoz-technology/mantil-cli/internal/terraform"
 	"github.com/atoz-technology/mantil-cli/pkg/mantil"
 	"github.com/spf13/cobra"
 )
@@ -16,7 +18,21 @@ var destroyCmd = &cobra.Command{
 	Args: cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		name := args[0]
-		os.RemoveAll(name)
+		_, err := os.Stat(name)
+		if err == nil {
+			fmt.Println("Destroying infrastructure...")
+			tf := terraform.New(name)
+			if err := tf.Init(); err != nil {
+				log.Fatal(err)
+			}
+			if err := tf.Plan(true); err != nil {
+				log.Fatal(err)
+			}
+			if err := tf.Apply(true); err != nil {
+				log.Fatal(err)
+			}
+			os.RemoveAll(name)
+		}
 		aws, err := aws.New()
 		if err != nil {
 			log.Fatal(err)
@@ -29,6 +45,7 @@ var destroyCmd = &cobra.Command{
 				log.Fatal(err)
 			}
 		}
+		fmt.Println("Deleting github repository...")
 		ghClient, err := github.NewClient()
 		if err != nil {
 			log.Fatal(err)
